+++ shared/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.1.4 (2024-09-24)
+
+- Change `milvusUrl` to `milvusUri` and `milvusApiKey` to `milvusToken`. 
+- Remove `milvusUsername` and `milvusPassword` as they were moved to the `milvusUri` as part of the URI.
+
 ## 0.1.3 (2024-09-11)
 
 - Set `performChunking` to `true` by default as some users were not aware of the setting their runs were failing.